+++ gettsim/taxes/lohn_st.py
@@ -125,6 +125,7 @@ def vorsorgepauschale_ab_2010(
     eink_st_abzuege_params:
       See params documentation :ref:`eink_st_abzuege_params`
 
+
     Returns
     -------
     Individual Vorsorgepauschale on annual basis
@@ -135,6 +136,7 @@ def vorsorgepauschale_ab_2010(
         12
         * (bruttolohn_m * soz_vers_beitr_params["beitr_satz"]["ges_rentenv"])
         * vorsorg_rv_anteil
+        * float(vorsorg_rv_anteil(eink_st_abzuege_params))
     )
 
     # 2. Krankenversicherungsbeiträge, §39b (2) Nr. 3b EStG.
@@ -191,7 +193,7 @@ def vorsorg_rv_anteil(eink_st_abzuege_params: dict, jahr: int):
     """
 
     out = piecewise_polynomial(
-        x=jahr,
+        x=jahr,      
         thresholds=eink_st_abzuege_params["vorsorge_pauschale_rv_anteil"]["thresholds"],
         rates=eink_st_abzuege_params["vorsorge_pauschale_rv_anteil"]["rates"],
         intercepts_at_lower_thresholds=eink_st_abzuege_params[
@@ -200,3 +202,78 @@ def vorsorg_rv_anteil(eink_st_abzuege_params: dict, jahr: int):
     )
 
     return out
+
+
+def steuerklasse(
+    tu_id: int,
+    gemeinsam_veranlagt: bool,
+    alleinerziehend: bool,
+    bruttolohn_m: float,
+    eink_st_params: dict,
+    eink_st_abzuege_params: dict,
+) -> int:
+    """Determine Lohnsteuerklassen (also called 'tax brackets')
+    They determine the basic allowance for the withdrawal tax
+
+    1: Single
+    2: Single Parent
+    3: One spouse in married couple who receives allowance of both partners.
+       Makes sense primarily for Single-Earner Households
+    4: Both spouses receive their individual allowance
+    5: If one spouse chooses 3, the other has to choose 5,
+       which means no allowance.
+    6: Additional Job...not modelled yet, as we do not
+    distinguish between different jobs
+
+    Parameters
+    ----------
+    tu_id: int
+        See basic input variable :ref:`tu_id <tu_id>`.
+    gemeinsam_veranlagt: bool
+        Return of :func:`gemeinsam_veranlagt`.
+    alleinerziehend: bool
+        See basic input variable :ref:`alleinerziehend <alleinerziehend>`.
+    bruttolohn_m: float
+        See basic input variable :ref:`bruttolohn_m <bruttolohn_m>`.
+    eink_st_params:
+        See params documentation :ref:`eink_st_params <eink_st_params>`
+    eink_st_abzuege_params:
+        See params documentation :ref:`eink_st_abzuege_params <eink_st_abzuege_params>`
+
+    Returns
+    ----------
+    steuerklasse: int
+        The steuerklasse for each person in the tax unit
+    """
+
+    bruttolohn_max = bruttolohn_m.groupby(tu_id).max()
+    bruttolohn_min = bruttolohn_m.groupby(tu_id).min()
+    # Determine Single Earner Couple:
+    # If one of the spouses earns tax-free income, assume Single Earner Couple
+    einkommensgrenze_zweitverdiener = (
+        eink_st_params["eink_st_tarif"]["thresholds"][1]
+        + eink_st_abzuege_params["werbungskostenpauschale"]
+    )
+    alleinverdiener_paar = (
+        (bruttolohn_min <= einkommensgrenze_zweitverdiener / 12)
+        & (bruttolohn_max > 0)
+        & (gemeinsam_veranlagt)
+    )
+    cond_steuerklasse1 = (~gemeinsam_veranlagt) & ~alleinerziehend
+    cond_steuerklasse2 = alleinerziehend
+    cond_steuerklasse3 = alleinverdiener_paar & (
+        bruttolohn_m > einkommensgrenze_zweitverdiener / 12
+    )
+    cond_steuerklasse4 = (gemeinsam_veranlagt) & (~alleinverdiener_paar)
+    cond_steuerklasse5 = alleinverdiener_paar & (
+        bruttolohn_m <= einkommensgrenze_zweitverdiener / 12
+    )
+    steuerklasse = (
+        1 * cond_steuerklasse1
+        + 2 * cond_steuerklasse2
+        + 3 * cond_steuerklasse3
+        + 4 * cond_steuerklasse4
+        + 5 * cond_steuerklasse5
+    )
+
+    return steuerklasse